[*]: *Massively* improve opacity computation time
A couple of major fixes, notably a new param to `Span.get_characters`
allows us to disregard the color part of a sequence, which we already
stored seperately per cell.

There is not higher-level opacity support yet so the performance can't
quite be tested, but at least in a case with opacity==1.0 performance
seems unaffected.

Beautiful.